BUILD 14
Changed the top of the normal obstacle, now it has a spike.
The revert portal now resets the status of the player and the enviroment
When the level is reloaded, all the obstacles in the scene will be destroyed
Fixed bug that spawns speed obstacle when the time  is already changed
When the jump animation ends, the player reset his scale.