BUILD 15
Added Pause Manager:
When the player press Esc or the pause button, the game will pause.
When paused the game and clicked on the restart button, the game will disable the player class for prevent the player jumping after unpause the game.

Remove Debug.Log on Spawner Class

Change the greound Sprite